Tweak test setup for download command
We're working on implementing teams.

On the implementation side we need to handle the "team" key, which
is either null or an object containing the name and slug of the team.

In the fake download server, we now pass back 'null' for the team by
default, and an object if there's a 'team' flag.

The contents of the solution file will change, so I moved this up
into the test setup, and moved the assertion out of the
assertDownloaded... function, since that function needs to be as stupid
as possible (otherwise it turns into information hiding, and that's kind
of disastrous).

The tests need multiple flags, so I switched to using a
map[flagName]flagValue{} in the test setup.

Lastly, the fake download server hard-codes files. We hard-code the assertions.
We were _not_ hard-coding the files in the fake response payload. This
changes it so that they're hard-coded there as well.